address input issues
