+++ lightly_studio_view/src/lib/components/ui/input/input.stories.svelte
@@ -1,6 +1,6 @@
 <script module>
     import { defineMeta } from '@storybook/addon-svelte-csf';
-    import Input from './Input.svelte';
+    import Input from './input.svelte';
 
     const { Story } = defineMeta({
         title: 'Components/Primitives/Input',
