+++ lightly_studio_view/src/lib/components/ui/input/input.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { render } from '@testing-library/svelte';
 import '@testing-library/jest-dom';
-import Input from './Input.svelte';
+import Input from './input.svelte';
 
 const progressSelector = '[data-testid="input-progress"]';
 const inputSelector = 'input';
